refactor(vm): refactor test + some linting
